+++ src/components/footer/footer.js
@@ -23,18 +23,21 @@ class Footer extends HTMLElement {
                     ${RenderService.showComponent(FooterBoard.contact.enabled, '<footer-contact></footer-contact>')}
                     ${RenderService.showComponent(FooterBoard.links.enabled, `   
                         <footer-link-section 
+                            id="quicklinks"
                             title="QuickLinks"
                             url="sfoxinsurance.org/quicklinks"
                             links='["Calculators", "Prices", "Terms of Service", "Privacy and security"]'
                         > 
                         </footer-link-section>
                         <footer-link-section
+                            id="join"
                             title="Join SFoxInsurance"
                             url="sfoxinsurance.org/join"
                             links='["Client programs", "Jobs", "Business customers", "Investments"]'
                         >
                         </footer-link-section>
                         <footer-link-section 
+                            id="tools"
                             title="Tools"
                             url="sfoxinsurance.org/tools"
                             links='["Everyday insurance", "Loans", "Insurance", "Cards", "Stocks"]'