+++ src/components/footer/link-section.js
@@ -1,6 +1,6 @@
 // @ts-nocheck
 import { CommonEvents } from '../../settings';
-import { ClassIdService, CustomEventService, StyleService } from '../../services';
+import { ClassIdService, CustomEventService, IdService, StyleService } from '../../services';
 import { ArrayService, JSONService, MobileService } from '../../services/utils';
 import { theme } from '../../theme/theme';
 
@@ -8,16 +8,16 @@ class FooterLinkSection extends HTMLElement {
     constructor() {
         super();
         this.shadow = this.attachShadow({mode: 'closed'});
+        this.id = this.getAttribute('id') || 'section';
         this.title = this.getAttribute('title') || 'Section';
         this.url = this.getAttribute('url') || 'sfoxinsurance.org';
 
-        CustomEventService.event(CommonEvents.click, this.toggleContent.bind(this), this.shadow);
         CustomEventService.event(CommonEvents.resize, this.updateSize.bind(this), window);
 
         const rowLinks = this.getAttribute('links');
         this.links = [];
         this.linksContent = '';
-        this.contentOpen = false;
+        this.contentOpen = true;
         this.screenW = window.innerWidth;
         const protocol = 'https://';
 
@@ -47,6 +47,7 @@ class FooterLinkSection extends HTMLElement {
     connectedCallback() {
         this.render();
         this.$content = ClassIdService.id('link-content', this.shadow);
+        this.$toggler = IdService.idAndClick(`toggler-${this.id}`, this.shadow, this.toggleContent.bind(this));
     }
 
     disconnectedCallback() {
@@ -127,7 +128,7 @@ class FooterLinkSection extends HTMLElement {
                 <div class="link-title">
                     ${this.title}
                     <div class="toggle-content">
-                        <div class="toggle-arrow" onclick="this.toggleContent()">
+                        <div id="toggler-${this.id}" class="toggle-arrow">
                             <svg
                                 width="16"
                                 height="16"