+++ Changelog.md
@@ -1,6 +1,7 @@
 ## Changelog
 v 2.51
-  Added improvement to fix lost focus of showModal, Added minLength improvements
+  Added improvement to fix lost focus of showModal, Added minLength improvements,
+  Added LinkSection toggle improvements (long awaited) :)
 
 v 2.50
   Improved Services utils to use classes, Added removeEvent and article paragraph